Start split of e2e tests
